Adds a TxBuild structure
