+++ lib/tx_build/default.ex
@@ -2,6 +2,8 @@ defmodule Stellar.TxBuild.Default do
   @moduledoc """
   Default TxBuild implementation.
   """
+  alias Stellar.TxBuild
+
   alias Stellar.TxBuild.{
     Account,
     Memo,
@@ -15,21 +17,9 @@ defmodule Stellar.TxBuild.Default do
 
   @behaviour Stellar.TxBuild.Spec
 
-  @type tx :: Transaction.t()
-  @type signatures :: Signature.t() | list(Signature.t())
-  @type tx_envelope :: TransactionEnvelope.t() | nil
-
-  @type t :: %__MODULE__{
-          tx: tx(),
-          signatures: signatures(),
-          tx_envelope: tx_envelope()
-        }
-
-  defstruct [:tx, :signatures, :tx_envelope]
-
   @impl true
   def new(%Account{} = account, opts \\ []) do
-    %__MODULE__{
+    %TxBuild{
       tx: Transaction.new(account, opts),
       signatures: [],
       tx_envelope: nil
@@ -37,52 +27,52 @@ defmodule Stellar.TxBuild.Default do
   end
 
   @impl true
-  def add_memo(%__MODULE__{tx: tx} = tx_build, %Memo{} = memo) do
+  def add_memo(%TxBuild{tx: tx} = tx_build, %Memo{} = memo) do
     transaction = %{tx | memo: memo}
     %{tx_build | tx: transaction}
   end
 
   @impl true
-  def set_timeout(%__MODULE__{tx: tx} = tx_build, %TimeBounds{} = timeout) do
+  def set_timeout(%TxBuild{tx: tx} = tx_build, %TimeBounds{} = timeout) do
     transaction = %{tx | time_bounds: timeout}
     %{tx_build | tx: transaction}
   end
 
   @impl true
-  def add_operation(%__MODULE__{} = tx_build, []), do: tx_build
+  def add_operation(%TxBuild{} = tx_build, []), do: tx_build
 
-  def add_operation(%__MODULE__{} = tx_build, [operation | operations]) do
+  def add_operation(%TxBuild{} = tx_build, [operation | operations]) do
     tx_build
     |> add_operation(operation)
     |> add_operation(operations)
   end
 
-  def add_operation(%__MODULE__{tx: tx} = tx_build, operation) do
+  def add_operation(%TxBuild{tx: tx} = tx_build, operation) do
     operation = Operation.new(operation)
     transaction = %{tx | operations: Operations.add(tx.operations, operation)}
     %{tx_build | tx: transaction}
   end
 
   @impl true
-  def sign(%__MODULE__{} = tx_build, []), do: tx_build
+  def sign(%TxBuild{} = tx_build, []), do: tx_build
 
-  def sign(%__MODULE__{} = tx_build, [%Signature{} = signature | signatures]) do
+  def sign(%TxBuild{} = tx_build, [%Signature{} = signature | signatures]) do
     tx_build
     |> sign(signature)
     |> sign(signatures)
   end
 
-  def sign(%__MODULE__{signatures: signatures} = tx_build, %Signature{} = signature) do
+  def sign(%TxBuild{signatures: signatures} = tx_build, %Signature{} = signature) do
     %{tx_build | signatures: signatures ++ [signature]}
   end
 
   @impl true
-  def build(%__MODULE__{tx: tx, signatures: signatures} = tx_build) do
+  def build(%TxBuild{tx: tx, signatures: signatures} = tx_build) do
     %{tx_build | tx_envelope: TransactionEnvelope.new(tx, signatures)}
   end
 
   @impl true
-  def envelope(%__MODULE__{tx: tx, signatures: signatures}) do
+  def envelope(%TxBuild{tx: tx, signatures: signatures}) do
     tx
     |> TransactionEnvelope.new(signatures)
     |> TransactionEnvelope.to_xdr()
